Remove PrimerLoadWebViewModelController, and move some of its logic in the root view controller. Should refactor that soon.
